Revision analyses
Analyses requested by the Reviewer. Rationale for the two-cohort analysis approach of the survey study. Clustering tendency. ROC analysis of subjective OD as a readout of objective hyposmia. Inverse distance-weighted kNN as a semi-supervised clustering approach. Multiple runs of permutation importance of the clustering factors. 'Kinetics' of objective OD in the CovILD study